Don't flatten volatility by default.
In previous versions, a large arbitrary scale was used to average volatility;
this had a side effect of very little detail being shown for small numbers or
small deltas.  Now the default is to keep the scale raw (1), but allow a
multiplier to flatten it out.  Also switching to Gemcutter releases with this
version.

* Support a second option to Bumpspark::Graph.new, `scale`. By default this is
  `1`